moving typheous connection stuff to separate module
